migrate native types with keeping native dir exports left
