+++ typings/tests/native-test.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
-import styled from "../../native";
+import styledFromNative from "../../native";
+import styled from "../..";
 
 const StyledView = styled.View`
   background-color: papayawhip;
@@ -10,6 +11,22 @@ const StyledText = styled.Text`
 `;
 
 class MyReactNativeComponent extends React.Component<{}, {}> {
+  render() {
+    return <StyledView>
+      <StyledText>Hello World!</StyledText>
+    </StyledView>;
+  }
+}
+
+const StyledViewFromNative = styledFromNative.View`
+  background-color: papayawhip;
+`;
+
+const StyledTextFromNative = styledFromNative.Text`
+  color: palevioletred;
+`;
+
+class MyReactNativeComponentFromNativeDir extends React.Component<{}, {}> {
   render() {
     return <StyledView>
       <StyledText>Hello World!</StyledText>
